게시글 수정

+++ lib/Community/Post_Widget.dart
@@ -25,7 +25,7 @@ class PostWidget extends StatefulWidget {
 class _PostWidgetState extends State<PostWidget> {
   late PostModel _model;
   bool _isModifying = false;
-
+  late TextEditingController _contentController;
   final scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
@@ -36,13 +36,42 @@ class _PostWidgetState extends State<PostWidget> {
     _model.textController ??= TextEditingController();
     _model.textFieldFocusNode ??= FocusNode();
     _model.getMemberIdx();
+    _contentController = TextEditingController();
   }
 
   @override
   void dispose() {
     _model.dispose();
-
+    _contentController.dispose();
     super.dispose();
+  }
+
+  void _toggleModifyingMode(PostElement post) {
+    setState(() {
+      _isModifying = !_isModifying;
+      if (_isModifying) {
+        _contentController.text = post.content ?? ''; // null 대신 빈 문자열을 사용하여 초기화합니다.
+      } else {
+        _contentController.clear();
+      }
+    });
+  }
+
+  Future<void> _saveChanges(PostElement post) async {
+    // 데이터 전송 전에 _contentController.text가 비어있지 않은지 확인합니다.
+    if (_contentController.text.isNotEmpty) {
+      // 내용이 있다면 업데이트를 진행합니다.
+      setState(() {
+        post.content = _contentController.text;
+        _isModifying = false;
+      });
+      // 여기서 서버에 데이터를 전송하는 로직을 추가합니다.
+    } else {
+      // 내용이 비어있다면 사용자에게 알립니다.
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text('내용을 입력해주세요.')),
+      );
+    }
   }
 
   @override
@@ -179,21 +208,17 @@ class _PostWidgetState extends State<PostWidget> {
                                 Spacer(), // Spacer 추가
                                 PopupMenuButton<String>(
                                   onSelected: (value) {
-                                    // if (value == 'modify') {
-                                    //   setState(() {
-                                    //     _isModifying = true;
-                                    //     _model.textController.text =
-                                    //         widget.univBoardId, context;
-                                    //   });
-                                    // } else if (value == 'delete') {
+                                    if (value == 'modify') {
+                                      _toggleModifyingMode(post);
+                                    } else if (value == 'delete') {
                                       _model
                                           .deletePost(
                                               widget.univBoardId, context)
                                           .then((_) {
                                         setState(() {});
                                       });
-                                    },
-                                
+                                    }
+                                  },
                                   itemBuilder: (BuildContext context) {
                                     return [
                                       PopupMenuItem(
@@ -208,7 +233,8 @@ class _PostWidgetState extends State<PostWidget> {
                                   },
                                   icon: Icon(
                                     Icons.more_vert,
-                                    color: FlutterFlowTheme.of(context).primaryText,
+                                    color: FlutterFlowTheme.of(context)
+                                        .primaryText,
                                   ),
                                 ),
                               ],
@@ -218,31 +244,69 @@ class _PostWidgetState extends State<PostWidget> {
                         alignment: AlignmentDirectional(-1, 0),
                         child: Padding(
                           padding: EdgeInsetsDirectional.fromSTEB(60, 15, 0, 0),
-                          child: AutoSizeText(
-                            post.content,
-                            style: FlutterFlowTheme.of(context)
-                                .bodyMedium
-                                .override(
-                                  fontFamily: 'Readex Pro',
-                                  fontSize: 18,
-                                  letterSpacing: 0,
-                                  fontWeight: FontWeight.normal,
+                          child: _isModifying
+                              ? TextFormField(
+                                  controller: _contentController,
+                                  maxLines: null,
+                                  decoration: InputDecoration(
+                                    hintText: '내용을 입력하세요...',
+                                    border: OutlineInputBorder(),
+                                  ),
+                                )
+                              : AutoSizeText(
+                                  post.content,
+                                  style: FlutterFlowTheme.of(context)
+                                      .bodyMedium
+                                      .override(
+                                        fontFamily: 'Readex Pro',
+                                        fontSize: 18,
+                                        letterSpacing: 0,
+                                        fontWeight: FontWeight.normal,
+                                      ),
                                 ),
-                          ),
                         ),
                       ),
+                      if (_isModifying)
+                        Align(
+                          alignment: AlignmentDirectional(0, 0),
+                          child: Padding(
+                            padding: EdgeInsetsDirectional.fromSTEB(60, 10, 0, 0),
+                            child: Row(
+                              children: [
+                                ElevatedButton(
+                                  onPressed: () async {
+                                    await _model.modifyPost(
+                                      post.univBoardId,
+                                      post.title,
+                                      _contentController.text,
+                                      post.categoryId!,
+                                      context,
+                                    );
+                                    _saveChanges(post);
+                                  },
+                                  child: Text('저장'),
+                                ),
+                                SizedBox(width: 10),
+                                ElevatedButton(
+                                  onPressed: () => _toggleModifyingMode(post),
+                                  child: Text('취소'),
+                                  style: ElevatedButton.styleFrom(
+                                    backgroundColor: Colors.red,
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
                       Padding(
                         padding: EdgeInsetsDirectional.fromSTEB(0, 10, 0, 20),
-                        child: post.postImageUrls
-                                .isEmpty // Check if the list is empty
-                            ? SizedBox() // Render an empty SizedBox if there are no images
+                        child: post.postImageUrls.isEmpty
+                            ? SizedBox()
                             : ClipRRect(
                                 borderRadius: BorderRadius.circular(8),
-                                child: post.postImageUrls.length ==
-                                        1 // Check if there's only one image
+                                child: post.postImageUrls.length == 1
                                     ? Image.network(
-                                        post.postImageUrls[
-                                            0], // Display the first image
+                                        post.postImageUrls[0],
                                         width:
                                             MediaQuery.sizeOf(context).width *
                                                 0.9,
@@ -251,8 +315,7 @@ class _PostWidgetState extends State<PostWidget> {
                                     : GridView.builder(
                                         gridDelegate:
                                             SliverGridDelegateWithFixedCrossAxisCount(
-                                          crossAxisCount:
-                                              3, // You can adjust the number of columns
+                                          crossAxisCount: 3,
                                         ),
                                         itemCount: post.postImageUrls.length,
                                         itemBuilder: (context, index) {
@@ -265,14 +328,13 @@ class _PostWidgetState extends State<PostWidget> {
                               ),
                       ),
                       Padding(
-                        padding: EdgeInsets.only(left: 15), // 왼쪽 마진 추가
+                        padding: EdgeInsets.only(left: 15),
                         child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start, // 왼쪽 정렬
+                          crossAxisAlignment: CrossAxisAlignment.start,
                           children: [
                             for (var reply in replies)
                               Padding(
-                                padding:
-                                    EdgeInsets.only(bottom: 15), // 아래 여백 추가
+                                padding: EdgeInsets.only(bottom: 15),
                                 child: ReplyWidget(reply: reply),
                               ),
                           ],
@@ -280,7 +342,6 @@ class _PostWidgetState extends State<PostWidget> {
                       ),
                       Builder(builder: (context) {
                         return Align(
-                          // 댓글입력
                           alignment: AlignmentDirectional(0, -1),
                           child: Row(
                             mainAxisSize: MainAxisSize.max,
@@ -429,7 +490,6 @@ class _PostWidgetState extends State<PostWidget> {
                           ),
                         );
                       }),
-                      // 기타 위젯 코드 생략
                     ],
                   ),
                 ),
